refactor: move package_order to rezplugin system
